+++ src/main/java/ddingdong/ddingdongBE/domain/notice/service/NoticeService.java
@@ -15,8 +15,8 @@ import ddingdong.ddingdongBE.domain.notice.entity.Notice;
 import ddingdong.ddingdongBE.domain.notice.repository.NoticeRepository;
 import ddingdong.ddingdongBE.domain.user.entity.User;
 import ddingdong.ddingdongBE.file.FileStore;
-import ddingdong.ddingdongBE.file.dto.FileResponse;
 
+import ddingdong.ddingdongBE.file.service.dto.FileResponse;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
